+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 - Support parsing empty subfield values
 - Set empty/zero occurrence to empty string instead of null
-- Add Stream parser
+- Add stream parsers